+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQXRQZ7TAFVTHH62YP0XHMTJ"
+        "01KQY7B5AHPE6DQMBT31S25V84"
     )
     r"""The unique identifier of the data source"""