+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQXRQZ7SMCA139W31M0TZTGK"
+        "01KQY7B5AG5CZR14DNK3HGW7BP"
     )
     r"""The unique identifier of the data source"""