+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQXRQZ7XQGTYCKGPRDKJPE5C"
+        "01KQY7B5AMD7FAJRH5W868VA4F"
     )
     r"""The unique identifier of the data source"""